#1972 - script validations added

+++ src/org/scada_lts/web/mvc/api/ScriptsAPI.java
@@ -3,9 +3,7 @@ package org.scada_lts.web.mvc.api;
 import br.org.scadabr.rt.scripting.ScriptRT;
 import br.org.scadabr.vo.scripting.ContextualizedScriptVO;
 import br.org.scadabr.vo.scripting.ScriptVO;
-import com.serotonin.db.IntValuePair;
 import com.serotonin.mango.Common;
-import com.serotonin.mango.vo.DataPointVO;
 import com.serotonin.mango.vo.User;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -20,6 +18,10 @@ import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import java.util.*;
+
+import static org.scada_lts.utils.MailingListApiUtils.isXidChanged;
+import static org.scada_lts.utils.ScriptsApiUtils.*;
+import static org.scada_lts.utils.ValidationUtils.formatErrorsJson;
 
 /**
  * Simple controller for Scripts in Scada-LTS
@@ -81,13 +83,17 @@ public class ScriptsAPI {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Integer> deleteScript(@PathVariable("id") int id, HttpServletRequest request) {
+    public ResponseEntity<String> deleteScript(@PathVariable Integer id, HttpServletRequest request) {
         LOG.info("DELETE::/api/scripts");
         try {
             User user = Common.getUser(request);
             if (user != null && user.isAdmin()) {
+                String error = validateScriptDelete(id);
+                if (!error.isEmpty()) {
+                    return ResponseEntity.badRequest().body(formatErrorsJson(error));
+                }
                 scriptService.deleteScript(id);
-                return new ResponseEntity<Integer>(id, HttpStatus.OK);
+                return new ResponseEntity<>(String.valueOf(id), HttpStatus.OK);
             } else {
                 return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
             }
@@ -97,14 +103,28 @@ public class ScriptsAPI {
     }
 
     @PostMapping(value = "/save")
-    public ResponseEntity<Map<String, Integer>> saveScript(@RequestBody JsonScript jsonBodyRequest, HttpServletRequest request) {
+    public ResponseEntity<Map<String, Object>> saveScript(@RequestBody JsonScript jsonBodyRequest, HttpServletRequest request) {
         LOG.info("POST::/api/scripts/save");
         try {
             User user = Common.getUser(request);
             if (user != null && user.isAdmin()) {
-                ContextualizedScriptVO vo = createScriptFromBody(jsonBodyRequest, user);
+                String error = validateScriptBody(jsonBodyRequest);
+                Map<String, Object> response = new HashMap<>();
+                if (!error.isEmpty()) {
+                    response.put("errors", error);
+                    return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+                }
+                if (isScriptPresent(jsonBodyRequest.getXid(), scriptService)) {
+                    response.put("errors", "This XID is already in use");
+                    return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+                }
+                String pointsError = validatePointsOnContext(jsonBodyRequest.getPointsOnContext(), dataPointService);
+                if (!pointsError.isEmpty()) {
+                    response.put("errors", pointsError);
+                    return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+                }
+                ContextualizedScriptVO vo = createScriptFromBody(jsonBodyRequest, user, dataPointService);
                 scriptService.saveScript(vo);
-                Map<String, Integer> response = new HashMap<>();
                 response.put("scriptId", vo.getId());
                 return new ResponseEntity<>(response, HttpStatus.CREATED);
             } else {
@@ -121,9 +141,11 @@ public class ScriptsAPI {
         try {
             User user = Common.getUser(request);
             if (user != null && user.isAdmin()) {
-                ContextualizedScriptVO vo = createScriptFromBody(jsonBodyRequest, user);
-                scriptService.saveScript(vo);
-                return new ResponseEntity<>("Script updated", HttpStatus.OK);
+                String error = validateScriptUpdate(jsonBodyRequest);
+                if (!error.isEmpty()) {
+                    return ResponseEntity.badRequest().body(formatErrorsJson(error));
+                }
+                return findAndUpdateScript(jsonBodyRequest);
             } else {
                 return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
             }
@@ -132,31 +154,36 @@ public class ScriptsAPI {
         }
     }
 
-    private ContextualizedScriptVO createScriptFromBody(JsonScript jsonBodyRequest, User user) {
-        ContextualizedScriptVO vo = new ContextualizedScriptVO();
-        vo.setId(jsonBodyRequest.getId());
-        vo.setXid(jsonBodyRequest.getXid());
-        vo.setName(jsonBodyRequest.getName());
-        vo.setScript(jsonBodyRequest.getScript());
-        vo.setPointsOnContext(convertPointsOnContext(jsonBodyRequest.getPointsOnContext()));
-        vo.setObjectsOnContext(convertObjectsOnContext(jsonBodyRequest));
-        vo.setUserId(user.getId());
-        return vo;
-    }
-
-    private List<IntValuePair> convertPointsOnContext(List<ScriptPoint> pointsOnContext) {
-        List<IntValuePair> points = new ArrayList<>();
-        for (ScriptPoint point : pointsOnContext) {
-            DataPointVO dp = dataPointService.getDataPoint(point.getDataPointXid());
-            points.add(new IntValuePair(dp.getId(), point.getVarName()));
+    @GetMapping(value = "/generateXid")
+    public ResponseEntity<String> getUniqueXid(HttpServletRequest request) {
+        try {
+            User user = Common.getUser(request);
+            if(user != null && user.isAdmin()) {
+                return new ResponseEntity<>(scriptService.generateUniqueXid(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return points;
     }
 
-    private List<IntValuePair> convertObjectsOnContext(JsonScript jsonBodyRequest) {
-        List<IntValuePair> objects = new ArrayList<>();
-        objects.add(new IntValuePair(1, jsonBodyRequest.getDatasourceContext()));
-        objects.add(new IntValuePair(2, jsonBodyRequest.getDatapointContext()));
-        return objects;
+    private ResponseEntity<String> findAndUpdateScript(JsonScript body) {
+        return getScript(body.getId(), scriptService).map(toUpdate -> updateScriptBody(toUpdate, body))
+                .orElse(new ResponseEntity<>(formatErrorsJson("Script not found"), HttpStatus.NOT_FOUND));
+    }
+
+    private ResponseEntity<String> updateScriptBody(ContextualizedScriptVO toUpdate, JsonScript body) {
+        if (isXidChanged(toUpdate.getXid(), body.getXid()) &&
+                isScriptPresent(body.getXid(), scriptService)){
+            return new ResponseEntity<>(formatErrorsJson("This XID is already in use"), HttpStatus.BAD_REQUEST);
+        }
+        String pointsError = validatePointsOnContext(body.getPointsOnContext(), dataPointService);
+        if (!pointsError.isEmpty()) {
+            return new ResponseEntity<>(formatErrorsJson(pointsError), HttpStatus.NOT_FOUND);
+        }
+        updateValueScript(toUpdate, body, dataPointService);
+        scriptService.saveScript(toUpdate);
+        return new ResponseEntity<>("{\"status\":\"updated\"}", HttpStatus.OK);
     }
 }
